Add code for extracting PDF patents from the OPS API.
PDFPatentPipeline is the entry point. The tool requires a JSON
configuration file (see below), a country code, and a year. An
optional argument allows a specific stage in the pipeline to be
targetted, and the system will automatically carry out all the
prerequisite stages. If no stage is specified, the tool will run all
the stages. The stages are

  search - find patent IDs for the country code and year
  biblio - find the date, download titles and abstracts
  fulltext - find whether claims/ description exist
  images - find whether PDF files exist
  claims - download the claims
  description - download the descriptions
  pdf - download the PDF files if text is missing (see below)

The tool attempts to extract all the text values - title, abstract,
claims, description - into tab-separated text files. If any of the
text parts is missing, it will also download PDF files with images of
the original patent publication, if available. Finding all four text
elements in any language will mean the PDF file is not downloaded.

If a text element has no language attribute, or it reports its
language using the code "ol", then the country code will be used
instead. Country codes are in upper case, while (genuine) language
codes are in lower case.

This code deals with the throttling limits set by the OPS API. Results
are saved periodically to checkpoint files to avoid duplication of
calls if the extraction process is stopped and restarted.

It uses the existing code in the utils directory for configuring the
system, obtaining an authentication token, calling the API with the
token, and logging. A suitable configuration file is needed, in the
same format as for the earlier code in tools and utils.